obs-be/requirements.txt
fastapi
uvicorn
clickhouse-connect
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
